NFC: moving code from bottom-up pass that belongs to SAS 17 algo
