+++ include/sea_dsa/BottomUp.hh
@@ -9,7 +9,6 @@
 
 #include "sea_dsa/CallSite.hh"
 #include "sea_dsa/Graph.hh"
-#include "sea_dsa/Mapper.hh"
 
 namespace llvm {
 class DataLayout;
@@ -27,23 +26,13 @@ public:
   typedef llvm::DenseMap<const llvm::Function *, GraphRef> GraphMap;
 
 private:
-  typedef std::shared_ptr<SimulationMapper> SimulationMapperRef;
-  typedef boost::container::flat_map<const llvm::Instruction *,
-                                     SimulationMapperRef>
-      CalleeCallerMapping;
 
   const llvm::DataLayout &m_dl;
   const llvm::TargetLibraryInfo &m_tli;
   const AllocWrapInfo &m_allocInfo;
   llvm::CallGraph &m_cg;
-  bool m_computeSimMap;
-  CalleeCallerMapping m_callee_caller_map;
-
   // true if assume that alloca allocated (stack) memory does not escape
   bool m_noescape;
-  // sanity check
-  bool checkAllNodesAreMapped(const llvm::Function &callee, Graph &calleeG,
-                              const SimulationMapper &sm);
 
 public:
 
@@ -53,23 +42,11 @@ public:
   BottomUpAnalysis(const llvm::DataLayout &dl,
                    const llvm::TargetLibraryInfo &tli,
                    const AllocWrapInfo &allocInfo, llvm::CallGraph &cg,
-		   bool computeSimMap,
                    bool noescape = true /* TODO: CLI*/)
       : m_dl(dl), m_tli(tli), m_allocInfo(allocInfo), m_cg(cg),
-	m_computeSimMap(computeSimMap), m_noescape(noescape) {}
+	m_noescape(noescape) {}
 
   bool runOnModule(llvm::Module &M, GraphMap &graphs);
-
-  typedef typename CalleeCallerMapping::const_iterator
-      callee_caller_mapping_const_iterator;
-
-  callee_caller_mapping_const_iterator callee_caller_mapping_begin() const {
-    return m_callee_caller_map.begin();
-  }
-
-  callee_caller_mapping_const_iterator callee_caller_mapping_end() const {
-    return m_callee_caller_map.end();
-  }
 };
 
 class BottomUp : public llvm::ModulePass {
