+++ src/DsaGlobal.cc
@@ -20,6 +20,7 @@
 #include "sea_dsa/Cloner.hh"
 #include "sea_dsa/Global.hh"
 #include "sea_dsa/Graph.hh"
+#include "sea_dsa/GraphUtils.hh"
 #include "sea_dsa/Local.hh"
 #include "sea_dsa/config.h"
 
@@ -42,8 +43,10 @@ using namespace llvm;
 
 namespace sea_dsa {
 
-/// CONTEXT-INSENSITIVE DSA
 
+//////
+/// Context-insensitive analysis
+/////
 // Unify callsite arguments within the same graph
 void ContextInsensitiveGlobalAnalysis::resolveArguments(DsaCallSite &cs, Graph &g) {
 
@@ -239,15 +242,40 @@ template <typename T> const T &WorkList<T>::dequeue() {
   return e;
 }
 
-/// CONTEXT-SENSITIVE DSA
+//////
+/// Context-sensitive analysis as described in SAS'17: bottom up +
+/// iterative (bottom-up/top-down) propagation on callsites.
+//////
+bool ContextSensitiveGlobalAnalysis::checkAllNodesAreMapped(
+     const Function &fn, Graph &fnG, const SimulationMapper &sm) {
+  std::set<const Node *> reach;
+  std::set<const Node *> retReach /*unused*/;
+  graph_utils::reachableNodes(fn, fnG, reach, retReach);
+  for (const Node *n : reach) {
+    Cell callerC = sm.get(Cell(const_cast<Node *>(n), 0));
+    if (callerC.isNull()) {
+      errs() << "ERROR: callee node " << *n
+             << " not mapped to a caller node.\n";
+      return false;
+    }
+  }
+  return true;
+}
 
 bool ContextSensitiveGlobalAnalysis::runOnModule(Module &M) {
 
+  
   LOG("dsa-global",
       errs() << "Started context-sensitive global analysis ... \n");
-
   // ufo::Stats::resume ("CS-DsaAnalysis");
-
+  
+  // Keep checks until implementation is stable
+#ifdef SANITY_CHECKS
+  const bool do_sanity_checks = true;
+#else
+  const bool do_sanity_checks = true;
+#endif   
+  
   for (auto &F : M) {
     if (F.isDeclaration() || F.empty())
       continue;
@@ -260,16 +288,68 @@ bool ContextSensitiveGlobalAnalysis::runOnModule(Module &M) {
   //    and initialize worklist
   BottomUpAnalysis bu(m_dl, m_tli, m_allocInfo, m_cg, true /*compute sim map*/);
   bu.runOnModule(M, m_graphs);
+  
+  // -- Compute simulation map so that we can identify which callsites
+  // -- require extra top-down propagation. Since bottom-up pass has
+  // -- been done already, we assume that the simulation relation is a
+  // -- total function (i.e., each callee node is mapped to a single
+  // -- node in the caller).  
+  CalleeCallerMapping callee_caller_map;
+  
+  for (auto it = scc_begin(&m_cg); !it.isAtEnd(); ++it) {
+    auto &scc = *it;
+    for (CallGraphNode *cgn : scc) {
+      Function *fn = cgn->getFunction();
+      if (!fn || fn->isDeclaration() || fn->empty()) {
+        continue;
+      }
+      // -- store the simulation maps from the SCC
+      for (auto &callRecord : *cgn) {
+	ImmutableCallSite CS(callRecord.first);
+	DsaCallSite dsaCS(CS);
+	const Function *callee = dsaCS.getCallee();
+	if (!callee || callee->isDeclaration() || callee->empty()) {
+	  continue;
+	}
+	
+	assert(m_graphs.count(dsaCS.getCaller()) > 0);
+	assert(m_graphs.count(dsaCS.getCallee()) > 0);
+	
+	Graph &callerG = *(m_graphs.find(dsaCS.getCaller())->second);
+	Graph &calleeG = *(m_graphs.find(dsaCS.getCallee())->second);
+	
+	SimulationMapperRef sm(new SimulationMapper());
+	bool res = Graph::computeCalleeCallerMapping(dsaCS, calleeG, callerG,
+						     *sm, do_sanity_checks);
+	if (!res) {
+	  llvm_unreachable("Simulation mapping check failed");
+	}
+	callee_caller_map.insert(std::make_pair(dsaCS.getInstruction(), sm));
+	
+	if (do_sanity_checks) {
+	  // Check the simulation map is a function
+	  if (!sm->isFunction()) {
+	    errs() << "ERROR (sea-dsa): simulation map for " << *dsaCS.getInstruction()
+		   << " is not a function!\n";
+	  } else {
+	    // Check the simulation map is a total function: check
+	    // that all nodes in the callee are mapped to one node in
+	    // the caller graph
+	    checkAllNodesAreMapped(*callee, calleeG, *sm);
+	  }
+	}
+      }
+    }
+  }
 
   DsaCallGraph dsaCG(m_cg);
   dsaCG.buildDependencies();
 
-  WorkList<const Instruction *> w;
-
   /// push in the worklist callsites for which two different
-  /// callee nodes are mapped to the same caller node
-  for (auto &kv : llvm::make_range(bu.callee_caller_mapping_begin(),
-                                   bu.callee_caller_mapping_end())) {
+  /// callee nodes are mapped to the same caller node  
+  WorkList<const Instruction *> w;
+  for (auto &kv : llvm::make_range(callee_caller_map.begin(),
+                                   callee_caller_map.end())) {
     auto const &simMapper = *(kv.second);
     assert(simMapper.isFunction());
 
@@ -336,11 +416,13 @@ bool ContextSensitiveGlobalAnalysis::runOnModule(Module &M) {
       errs() << "-- Number of top-down propagations=" << td_props << "\n";
       errs() << "-- Number of bottom-up propagations=" << bu_props << "\n";);
 
-#ifdef SANITY_CHECKS
-  assert(checkNoMorePropagation());
-#endif
+  if (do_sanity_checks) {
+    if (!checkNoMorePropagation()){
+      errs() << "ERROR (sea-dsa) sanity check failed: more top-down or "
+	     << "bottom-up propagation is needed\n";
+    }
+  }
 
-  /// FIXME: propagate both in the same fixpoint
   if (normalizeUniqueScalars) {
     CallGraphClosure<ContextSensitiveGlobalAnalysis, UniqueScalar> usa(*this,
                                                                        dsaCG);
@@ -441,7 +523,7 @@ bool ContextSensitiveGlobalAnalysis::checkNoMorePropagation() {
         PropagationKind pkind = decidePropagation(cs, calleeG, callerG);
         if (pkind != NONE) {
           auto pkind_str = (pkind == UP) ? "bottom-up" : "top-down";
-          errs() << "ERROR sanity check failed:" << *(cs.getInstruction())
+          errs() << "ERROR (sea-dsa) sanity check failed:" << *(cs.getInstruction())
                  << " requires " << pkind_str << " propagation.\n";
           return false;
         }
@@ -465,6 +547,9 @@ bool ContextSensitiveGlobalAnalysis::hasGraph(const Function &fn) const {
   return m_graphs.count(&fn) > 0;
 }
 
+///////
+/// Context-sensitive analysis: bottom-up + top-down
+///////
 bool BottomUpTopDownGlobalAnalysis::runOnModule(Module &M) {
   LOG("dsa-global",
       errs() << "Started bottom-up + top-down global analysis ... \n");
